test(FakeMovieData): 테스트를 위한 FakeMovie 생성
